+++ evaldo/builtins.go
@@ -2277,53 +2277,6 @@ var builtins = map[string]*env.Builtin{
 		},
 	},
 
-	"enter-console": {
-		Argsn: 1,
-		Doc:   "Stops execution and gives you a Rye console, to test the code inside environment.",
-		Fn: func(ps *env.ProgramState, arg0 env.Object, arg1 env.Object, arg2 env.Object, arg3 env.Object, arg4 env.Object) env.Object {
-			switch name := arg0.(type) {
-			case env.String:
-				ser := ps.Ser
-				/* ps.Ser = bloc.Series
-				EvalBlock(ps)
-				ps.Ser = ser */
-				//reader := bufio.NewReader(os.Stdin)
-
-				fmt.Println("Welcome to console: \033[1m" + name.Value + "\033[0m")
-				fmt.Println("* use \033[1mls\033[0m to list current context")
-				fmt.Println("-------------------------------------------------------------")
-				/*
-					for {
-						fmt.Print("{ rye dropin }")
-						text, _ := reader.ReadString('\n')
-						//fmt.Println(1111)
-						// convert CRLF to LF
-						text = strings.Replace(text, "\n", "", -1)
-						//fmt.Println(1111)
-						if strings.Compare("(lc)", text) == 0 {
-							fmt.Println(ps.Ctx.Print(*ps.Idx))
-						} else if strings.Compare("(r)", text) == 0 {
-							ps.Ser = ser
-							return ps.Res
-						} else {
-							// fmt.Println(1111)
-							block, genv := loader.LoadString("{ " + text + " }")
-							ps := env.AddToProgramState(ps, block.Series, genv)
-							EvalBlock(ps)
-							fmt.Println(ps.Res.Inspect(*ps.Idx))
-						}
-					}*/
-
-				DoRyeRepl(ps, "do", ShowResults)
-				fmt.Println("-------------------------------------------------------------")
-				ps.Ser = ser
-				return ps.Res
-			default:
-				return MakeArgError(ps, 1, []env.Type{env.StringType}, "enter-console")
-			}
-		},
-	},
-
 	/* TEMP FOR WASM 20250116
 	"get-input": {
 		Argsn: 1,
@@ -7831,6 +7784,7 @@ func RegisterBuiltins(ps *env.ProgramState) {
 	RegisterBuiltins2(Builtins_smtpd, ps, "smtpd")
 	RegisterBuiltins2(Builtins_mail, ps, "mail")
 	RegisterBuiltins2(Builtins_ssh, ps, "ssh")
+	RegisterBuiltins2(Builtins_console, ps, "console")
 	RegisterBuiltinsInContext(Builtins_math, ps, "math")
 	RegisterBuiltinsInContext(Builtins_devops, ps, "devops")
 	// ## Archived modules
