+++ evaldo/builtins_console_not.go
@@ -1,0 +1,7 @@
+//go:build b_norepl || wasm || js
+
+package evaldo
+
+import "github.com/refaktor/rye/env"
+
+var Builtins_console = map[string]*env.Builtin{}
