+++ runner/runner.go
@@ -1,3 +1,5 @@
+//go:build !wasm
+
 package runner
 
 import (
